Reorganized base record generation
More refactoring is needed. Currently, we use print-to-e default values, but these values in reducers should be taken by transformation type.
